requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.17
langchain-groq==0.0.1
langchain-pinecone==0.0.2
langgraph==0.0.20
pinecone-client==3.0.2
python-dotenv==1.0.1
polygon-api-client==1.13.3
tavily-python==0.3.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
streamlit==1.31.1
yfinance==0.2.36
pandas==2.2.0
ta==0.11.0
-e .